pet: cast scatter index to int64 in compute_batch_tensors
metatensor's neighbor-list samples are int32 (`Labels` are stored int32 in
the Rust core), but torch.scatter_add_'s index argument must be int64.
Blocked any real forward pass on this branch with:

    RuntimeError: scatter(): Expected dtype int64 for index

Reproduced identically on both L40S and H100, with both the default and a
much larger PET model, so it's unconditional -- not a size- or
device-dependent issue. Likely never hit before because older torch
versions validated scatter index dtypes less strictly.

Found running pipeline-bench (a Snakemake harness benchmarking PET training
across metatrain PR branches) as a real GPU smoke test on kuma.

+++ src/metatrain/pet/modules/structures.py
@@ -284,7 +284,7 @@ def compute_batch_tensors(
     # static (backed) shape.
     num_neighbors = torch.zeros(
         num_nodes, dtype=centers.dtype, device=centers.device
-    ).scatter_add_(0, centers, torch.ones_like(centers))
+    ).scatter_add_(0, centers.long(), torch.ones_like(centers))
     # ``max_edges_per_node`` (the largest neighbour count of any atom) becomes the size
     # of the NEF grid's second dimension. The ``numel`` guard keeps empty systems
     # (no atoms) well defined.
